Make game clear
