+++ Resources/stages/stage01
@@ -5,3 +5,8 @@
 1 1 1 0 1 1 1
 1 0 0 0 2 5 1
 1 1 1 1 1 1 1
+4
+6 4 1
+7 5 1
+8 2 3
+9 1 3
