+++ stage.md
@@ -12,6 +12,10 @@ row                     (縦の長さ)
 . . .
 . . .
 ブロックを表す識別子        (row 行 * col列)
+goalの個数
+識別子 x y
+ :
+識別子 x y
 ```
 カッコ内は記述しない
